#28 Updated Docker Compose setup.
I added in 2 sub files, with settings for debug and settings for release. Those files are automatically detected by visual studio's configuration manager which allows for seamless integration. 

This commit is necessary so that next sprint when we need to host our site, we can create a production build of it without needing to manually change most settings. In the future if needed, more configuration settings can be added to the specific debug & release files.

Breaking changes: when running in release mode, you'll need to include our self signed certificate. Details can be found in our team chat. The .gitignore is set to not include certificates inside repositories for safety reasons.  When actually hosting on a domain our certificate will be different anyways.