#37 Created SignalR Hub for Client Server Game data exchange
This commit, establishes the hub and assigns it the route hub/gamedata . There is 4 potential actions the client can send through the hub, and currently the server just echo's the message to all clients. This will eventually send data to the interpolator that is currently being developed by another student.