#37 Added Tests for the GameData Hub. Renamed Position method
Using a new method of testing signalr hubs.  The hubs are tested using this signalr unit testing support library that allows us to mock up chat hubs and have them execute as it they are running on a server. The library provides a bunch of asserts which allows us to test the signalr hub. 

The tests currently reflect the progress made in the gamedata hub. Next task related to this will implement the changes added by the interpolator work.