Separated Phaser logic from React Component Logic
The way the game was setup, react's props were not accessible in the phaser update methods. In order to resolve this, I needed to extend the Phaser classes in order to pass down react props. In doing so, I've cleaned up the code by putting phaser logic in their own classes. I also separated the Phaser.Game from the Phaser.Scene. From what I've seen online this is best practice.

**Note The GameStartRenderer.tsx file has been deleted in favor for GameStartRenderer2.tsx. The name of the new file, can be updated later, the 2 is to provide emphasis that this was a major change