#67 Made auth-token cookie accessible from javascript
Without making significant overhauls to our current authentication scheme, in order to authenticate users over websockets we need to send the auth token in the body of the message. This requires us to be able to access the token from javascript. In the future if time allocates, we should try and come up with a better solution.