#67 Multiplayer Setup!!!! Made initialization of singlar async.
In order to setup multiplayer, the scene had to be overhauled. The scene's create method and update method now iterate through all the players in the playerdict. new players are added to the screen and old one's have their coordinates updated. 
In order to safely start the scene, the initialization of the websocket connection to the server was made async. Now the code will await for the connection to be established. This is why the globalchat store also had to be updated.