#75 Made GameLogic more concurrency friendly and fixed diagonal movement
I.e. when you hold down 2 keys at once you won't move in that direction for an inordinate amount of time. Also realized that there has to be a conversion factor between the units that the physics engine uses and the units we use to represent things in the game world.