#91Implemented frontend logic for sending and viewing projectiles
I updated the gamedata redux store to include the sending projectile action. I've also updated the interfaces to support the new types made for projectiles. I updated the create and update methods in the phaser scene to render the bullets that are in the game

I also included a temporary 'asset' for the projectiles so that they can be seen.